--- a/crypto/secp256k1/libsecp256k1/TODO
+++ b/crypto/secp256k1/libsecp256k1/TODO
@@ -1,3 +1,0 @@
-* Unit tests for fieldelem/groupelem, including ones intended to
-  trigger fieldelem's boundary cases.
-* Complete constant-time operations for signing/keygen