--- a/crypto/secp256k1/libsecp256k1/cmake/arm-linux-gnueabihf.toolchain.cmake
+++ b/crypto/secp256k1/libsecp256k1/cmake/arm-linux-gnueabihf.toolchain.cmake
@@ -1,0 +1,3 @@
+set(CMAKE_SYSTEM_NAME Linux)
+set(CMAKE_SYSTEM_PROCESSOR arm)
+set(CMAKE_C_COMPILER arm-linux-gnueabihf-gcc)