… crypto/secp256k1/libsecp256k1/.cirrus.yml
… crypto/secp256k1/libsecp256k1/.github/actions/install-homebrew-valgrind/action.yml
… crypto/secp256k1/libsecp256k1/.github/actions/run-in-docker-action/action.yml
… crypto/secp256k1/libsecp256k1/.github/workflows/ci.yml
… crypto/secp256k1/libsecp256k1/.travis.yml
… crypto/secp256k1/libsecp256k1/CHANGELOG.md
… crypto/secp256k1/libsecp256k1/CMakeLists.txt
… crypto/secp256k1/libsecp256k1/CMakePresets.json
… crypto/secp256k1/libsecp256k1/CONTRIBUTING.md
… crypto/secp256k1/libsecp256k1/README.md
… crypto/secp256k1/libsecp256k1/SECURITY.md
… crypto/secp256k1/libsecp256k1/build-aux/m4/ax_jni_include_dir.m4
… crypto/secp256k1/libsecp256k1/build-aux/m4/ax_prog_cc_for_build.m4
… crypto/secp256k1/libsecp256k1/ci/linux-debian.Dockerfile
… crypto/secp256k1/libsecp256k1/cmake/CheckStringOptionValue.cmake
… crypto/secp256k1/libsecp256k1/cmake/source_arm32.s
… crypto/secp256k1/libsecp256k1/cmake/x86_64-w64-mingw32.toolchain.cmake
… crypto/secp256k1/libsecp256k1/configure.ac
… crypto/secp256k1/libsecp256k1/contrib/dummy.go
… crypto/secp256k1/libsecp256k1/contrib/lax_der_parsing.c
… crypto/secp256k1/libsecp256k1/contrib/lax_der_parsing.h
… crypto/secp256k1/libsecp256k1/contrib/lax_der_privatekey_parsing.c
… crypto/secp256k1/libsecp256k1/contrib/lax_der_privatekey_parsing.h
… crypto/secp256k1/libsecp256k1/doc/ellswift.md
… crypto/secp256k1/libsecp256k1/doc/musig.md
… crypto/secp256k1/libsecp256k1/doc/release-process.md
… crypto/secp256k1/libsecp256k1/doc/safegcd_implementation.md
… crypto/secp256k1/libsecp256k1/examples/CMakeLists.txt
… crypto/secp256k1/libsecp256k1/examples/EXAMPLES_COPYING
… crypto/secp256k1/libsecp256k1/examples/ecdh.c