@MariusVanDerWijden 2025-02-24
It would be nice to sanity check this on all platforms